# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lexiflow API
  description: |
    REST API for the LingQ Clone language-learning app: authentication, library/courses,
    lesson reader, vocabulary, phrases, uploads, and optional LingQ integration.

    **Authentication:** Most routes require a JWT in the `Authorization` header as `Bearer <token>`.
    Obtain a token from `POST /auth/login` or `POST /auth/register` (register does not return a token; log in after sign-up).
  version: 1.0.0
  license:
    name: ISC
    identifier: ISC
servers:
  - url: http://localhost:3000/api/v1
    description: Local or Docker (default port 3000)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/login (`token` field)
  schemas:
...
    UserSession:
      type: object
      description: |
        Authenticated user profile fields returned from the database.
        Same shape from `POST /auth/login` and `GET /auth/verify`.
      required: [id, username, email]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
          description: Display name; may be an empty string if omitted at registration.
        preferences:
          type: object
          nullable: true
          properties:
            targetLanguage:
              type: string
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password, targetLanguage]
              properties:
                email:
                  type: string
                  format: email
                username:
                  type: string
                fullName:
                  type: string
                password:
                  type: string
                  format: password
                targetLanguage:
                  type: string
                  description: Language code (must exist in `languages` table)
                dailyGoalTier:
                  type: string
                  description: e.g. calm (default)
      responses:
        '200':
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Validation or unsupported language
        '409':
          description: Email or username already taken
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: JWT and user payload
          content:
            application/json:
              schema:
                type: object
                required: [token, user]
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/UserSession'
        '400':
          description: Invalid password
        '404':
          description: User not found
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out (client should discard token)
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Acknowledged
        '401':
          description: Missing or invalid token
  /auth/verify:
    get:
      tags: [Auth]
      summary: Verify JWT and load current user from the database
      description: |
        Validates the bearer token, then returns the same `user` object shape as `POST /auth/login`
        (`id`, `username`, `email`, `fullName`, `preferences`). Use this on app load to refresh
        the client session (not only the JWT subject `id`).
      operationId: verifyToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/UserSession'
        '401':
          description: Access denied or unauthorized
        '404':
          description: User id in token no longer exists in the database
        '500':
          description: Server error
  /auth/info/{userId}:
    get:
      tags: [Auth]
      summary: Aggregated profile and stats for a user
      description: |
        Public route (no JWT). Supports optional `lang` query to override target language.
        Returns streaks, daily totals, 7d/30d charts, RTL flag, etc.
      operationId: getUserInfo
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: lang
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Profile snapshot
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid user id or error
  /auth/goal-tier:
    patch:
      tags: [Auth]
      summary: Update daily goal tier for a language
      operationId: patchGoalTier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier, languageCode]
              properties:
                tier:
                  type: string
                languageCode:
                  type: string
      responses:
        '200':
          description: Updated
        '400':
          description: Missing fields
        '401':
          description: Unauthorized
  /auth/preferences:
    patch:
      tags: [Auth]
      summary: Set user target language preference
      operationId: patchPreferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetLanguage]
              properties:
                targetLanguage:
                  type: string
      responses:
        '200':
          description: Updated
        '401':
          description: Unauthorized
...
  /auth/profile/insights/{userId}:
    get:
      tags: [Auth]
      summary: Markov / learning analytics for profile (vocab + phrases)
      operationId: getProfileInsights
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vocab and phrase insight payloads
        '401':
          description: Unauthorized
  /auth/profile/reset/{languageCode}:
    post:
      tags: [Auth]
      summary: Reset all user data for a language (destructive)
      operationId: resetProfileLanguage
      security:
        - bearerAuth: []
      parameters:
        - name: languageCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reset complete
        '401':
          description: Unauthorized
...